requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
decorator==4.3.0
diff-match-patch==20121119
dj-pagination==2.4.0
Django==1.11.16
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.1.0
django-js-asset==1.1.0
django-ranged-response==0.2.0
django-reversion==3.0.0
django-simple-captcha==0.5.9
djangorestframework==3.9.0
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.7
ipython==7.0.1
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.13.1
Markdown==3.0.1
odfpy==1.3.6
openpyxl==2.5.9
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.5
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.2.0
PyMySQL==0.9.2
pytz==2018.5
PyYAML==3.13
requests==2.20.0
simplegeneric==0.8.1
six==1.10.0
tablib==0.12.1
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.24
wcwidth==0.1.7
xadmin==0.6.1
xlrd==1.1.0
xlwt==1.3.0
